refactor(full-offline): move logic to a service
